CXP-723: Extract scope list and some interfaces to ease EE work
