+++ src/Akeneo/Connectivity/Connection/back/Infrastructure/Symfony/Resources/translations/jsmessages.en_US.yml
@@ -63,34 +63,36 @@ akeneo_connectivity.connection:
             unreachable: We can't reach the marketplace, please come back later.
             scroll_to_top: Scroll to the top
         apps:
+            wizard:
+                title: Connect
+                action:
+                    confirm: Confirm
+                    cancel: Cancel
+                authorize:
+                    title: 'App {{ app_name }} needs to'
+                    no_scope_title: 'The {{ app_name }} App would like to access your PIM.'
+                    no_scope: No specific authorizations have been requested.
+                    helper: To know more about about app authorization,
+                    helper_link: check out our Help Center for more information.
             title: Connect
-            action:
-                confirm: Confirm
-                cancel: Cancel
             activate:
                 flash:
                     error: Sorry, an error occurred while connecting the App.
-            authorize:
-                title: 'App {{ app_name }} needs to'
-                no_scope_title: 'The {{ app_name }} App would like to access your PIM.'
-                no_scope: No specific authorizations have been requested.
-                helper: To know more about about app authorization,
-                helper_link: check out our Help Center for more information.
-                error:
-                    sub_text: Please contact the app developer or Akeneo support
-                scope:
-                    type:
-                        view: 'View {{ entities }}'
-                        edit: 'View and edit {{ entities }}'
-                        delete: 'View, edit and delete {{ entities }}'
-                    entities:
-                        catalog_structure: attributes, attribute groups, families and family variants
-                        attribute_options: attribute options
-                        categories: categories
-                        channel_localization: locales and currencies
-                        channel_settings: channels
-                        association_types: association types
-                        products: products and product models
+            scope:
+                type:
+                    view: 'View {{ entities }}'
+                    edit: 'View and edit {{ entities }}'
+                    delete: 'View, edit and delete {{ entities }}'
+                entities:
+                    catalog_structure: attributes, attribute groups, families and family variants
+                    attribute_options: attribute options
+                    categories: categories
+                    channel_localization: locales and currencies
+                    channel_settings: channels
+                    association_types: association types
+                    products: products and product models
+            error:
+                sub_text: Please contact the app developer or Akeneo support
             constraint:
                 client_id:
                     not_blank: The client_id is required.
