+++ src/Akeneo/Connectivity/Connection/front/src/connect/components/AppWizard/AppWizard.tsx
@@ -2,9 +2,11 @@ import React, {FC, useEffect, useState} from 'react';
 import styled from 'styled-components';
 import {Button, getColor, getFontSize, Modal} from 'akeneo-design-system';
 import {useHistory} from 'react-router';
-import {AppWizardData, useFetchAppWizardData} from '../../hooks/use-fetch-app-wizard-data';
+import {useFetchAppWizardData} from '../../hooks/use-fetch-app-wizard-data';
 import {useTranslate} from '../../../shared/translate';
-import {ScopeList} from './ScopeList';
+import {AppWizardData} from '../../../model/Apps/wizard-data';
+import {ScopeListContainer} from './ScopeListContainer';
+import {Authorizations} from "./Authorizations";
 
 const Content = styled.div`
     display: grid;
@@ -76,7 +78,7 @@ export const AppWizard: FC<Props> = ({clientId}) => {
     return (
         <Modal
             onClose={redirectToMarketplace}
-            closeTitle={translate('akeneo_connectivity.connection.connect.apps.action.cancel')}
+            closeTitle={translate('akeneo_connectivity.connection.connect.apps.wizard.action.cancel')}
         >
             <Content>
                 <LogoContainer>
@@ -84,18 +86,16 @@ export const AppWizard: FC<Props> = ({clientId}) => {
                         <Logo src={wizardData.appLogo} alt={wizardData.appName} />
                     </LogoFrame>
                 </LogoContainer>
-                <InfoContainer>
-                    <Connect>{translate('akeneo_connectivity.connection.connect.apps.title')}</Connect>
-                    <ScopeList appName={wizardData.appName} scopeMessages={wizardData.scopeMessages} />
+                <Authorizations appName={wizardData.appName} scopeMessages={wizardData.scopeMessages}>
                     <Actions>
                         <ActionButton level={'tertiary'} onClick={redirectToMarketplace}>
-                            {translate('akeneo_connectivity.connection.connect.apps.action.cancel')}
+                            {translate('akeneo_connectivity.connection.connect.apps.wizard.action.cancel')}
                         </ActionButton>
                         <ActionButton>
-                            {translate('akeneo_connectivity.connection.connect.apps.action.confirm')}
+                            {translate('akeneo_connectivity.connection.connect.apps.wizard.action.confirm')}
                         </ActionButton>
                     </Actions>
-                </InfoContainer>
+                </Authorizations>
             </Content>
         </Modal>
     );
