+++ src/Akeneo/Connectivity/Connection/front/src/connect/hooks/use-fetch-app-wizard-data.ts
@@ -1,18 +1,6 @@
 import {useRoute} from '../../shared/router';
 import {useCallback} from 'react';
 
-export interface ScopeMessage {
-    icon: string;
-    type: string;
-    entities: string;
-}
-
-export interface AppWizardData {
-    appName: string;
-    appLogo: string;
-    scopeMessages: ScopeMessage[];
-}
-
 export const useFetchAppWizardData = (clientId: string) => {
     const url = useRoute('akeneo_connectivity_connection_apps_rest_get_wizard_data', {clientId: clientId});
