+++ src/Akeneo/Connectivity/Connection/front/src/model/Apps/scope-message.ts
@@ -1,0 +1,5 @@
+export default interface ScopeMessage {
+    icon: string;
+    type: string;
+    entities: string;
+}
