+++ src/Akeneo/Connectivity/Connection/front/src/model/Apps/wizard-data.ts
@@ -1,0 +1,7 @@
+import ScopeMessage from './scope-message';
+
+export interface AppWizardData {
+    appName: string;
+    appLogo: string;
+    scopeMessages: ScopeMessage[];
+}
